If you have edited a message and then deleted it offline, when the connection is restored, the message is deleted

+++ app/src/main/java/mega/privacy/android/app/lollipop/controllers/ChatController.java
@@ -296,7 +296,8 @@ public class ChatController {
         }
 
         logDebug("Delete normal message with status = "+message.getStatus());
-        if(message.getStatus() == MegaChatMessage.STATUS_SENDING){
+        if(message.getStatus() == MegaChatMessage.STATUS_SENDING && message.getMsgId() == megaApi.INVALID_HANDLE){
+
             messageToDelete = megaChatApi.deleteMessage(chatId, message.getTempId());
         }else{
             messageToDelete = megaChatApi.deleteMessage(chatId, message.getMsgId());
